# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title; it controls generated filenames.
  title: Api
  version: 0.1.0
  description: Public shipment tracking and logistics operations API.
servers:
  - url: /api
...
paths:
...
  /admin/dashboard:
    get:
      operationId: getAdminDashboard
      tags: [admin]
      summary: Get live operational counts and recent tracking activity
      security: [{ clerkSession: [] }]
      responses:
        "200":
          description: Dashboard calculated from saved records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
  /admin/shipments:
    get:
      operationId: getAdminShipments
      tags: [admin]
      summary: Search and filter shipments
      security: [{ clerkSession: [] }]
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: destination
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
      responses:
        "200":
          description: Paginated shipments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShipmentPage"
    post:
      operationId: createShipment
      tags: [admin]
      summary: Create shipment with a unique internal tracking number
      security: [{ clerkSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShipmentInput"
      responses:
        "201":
          description: Created shipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"
  /admin/shipments/{id}:
    get:
      operationId: getAdminShipment
      tags: [admin]
      summary: Get shipment details and full event history
      security: [{ clerkSession: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Shipment and events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShipmentDetail"
    patch:
      operationId: updateShipment
      tags: [admin]
      summary: Update shipment details
      security: [{ clerkSession: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShipmentUpdate"
      responses:
        "200":
          description: Updated shipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"
  /admin/shipments/{id}/events:
    post:
      operationId: createTrackingEvent
      tags: [admin]
      summary: Add an internal logistics event
      security: [{ clerkSession: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackingEventInput"
      responses:
        "201":
          description: Added event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackingEvent"
  /admin/shipments/{id}/sync:
    post:
      operationId: syncShipmentTracking
      tags: [admin]
      summary: Register and synchronize with the configured carrier tracking provider
      security: [{ clerkSession: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Synchronization result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResult"
        "503":
          description: Provider is not configured or is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/quotes:
    get:
      operationId: getAdminQuotes
      tags: [admin]
      summary: Search and filter quote requests
      security: [{ clerkSession: [] }]
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Quote requests
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Quote" }
  /admin/quotes/{id}:
    patch:
      operationId: updateQuote
      tags: [admin]
      summary: Update quote status and private notes
      security: [{ clerkSession: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteUpdate"
      responses:
        "200":
          description: Updated quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
  /admin/customers:
    get:
      operationId: getAdminCustomers
      tags: [admin]
      summary: Search customer accounts
      security: [{ clerkSession: [] }]
      parameters:
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Customers
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/User" }
  /admin/contacts:
    get:
      operationId: getAdminContacts
      tags: [admin]
      summary: List contact inquiries
      security: [{ clerkSession: [] }]
      responses:
        "200":
          description: Inquiries
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Contact" }
  /admin/contacts/{id}:
    patch:
      operationId: updateContact
      tags: [admin]
      summary: Update contact inquiry status
      security: [{ clerkSession: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactUpdate"
      responses:
        "200":
          description: Updated inquiry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /admin/users:
    get:
      operationId: getAdminUsers
      tags: [admin]
      summary: List user roles and accounts
      security: [{ clerkSession: [] }]
      responses:
        "200":
          description: User accounts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/User" }
  /admin/users/{id}/role:
    patch:
      operationId: updateUserRole
      tags: [admin]
      summary: Change a user role (super admin only)
      security: [{ clerkSession: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/audit:
    get:
      operationId: getAdminAuditLogs
      tags: [admin]
      summary: View administrative audit records
      security: [{ clerkSession: [] }]
      responses:
        "200":
          description: Audit records
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AuditLog" }
  /admin/site:
    patch:
      operationId: updateSiteSettings
      tags: [admin]
      summary: Update configurable company website content
      security: [{ clerkSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteSettingsUpdate"
      responses:
        "200":
          description: Updated site settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"
components:
  securitySchemes:
    clerkSession:
      type: http
      scheme: bearer
      description: Authenticated Clerk browser session (same-origin cookie).
  schemas:
...
    Shipment:
      type: object
      required: [id, trackingNumber, serviceType, origin, destination, status, isDemo, createdAt, updatedAt]
      properties:
        id: { type: integer }
        trackingNumber: { type: string }
        customerId: { type: ["integer", "null"] }
        customerName: { type: ["string", "null"] }
        carrier: { type: ["string", "null"] }
        carrierTrackingNumber: { type: ["string", "null"] }
        serviceType: { type: string }
        origin: { type: string }
        destination: { type: string }
        currentLocation: { type: ["string", "null"] }
        status: { type: string }
        weight: { type: ["number", "null"] }
        dimensions: { type: ["string", "null"] }
        packageDescription: { type: ["string", "null"] }
        estimatedDelivery: { type: ["string", "null"], format: date }
        actualDelivery: { type: ["string", "null"], format: date }
        isDemo: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ShipmentDetail:
      type: object
      required: [shipment, events]
      properties:
        shipment: { $ref: "#/components/schemas/Shipment" }
        events:
          type: array
          items: { $ref: "#/components/schemas/TrackingEvent" }
...
    ShipmentInput:
      type: object
      required: [serviceType, origin, destination]
      properties:
        customerId: { type: ["integer", "null"] }
        carrier: { type: ["string", "null"], maxLength: 100 }
        carrierTrackingNumber: { type: ["string", "null"], maxLength: 100 }
        serviceType: { type: string, minLength: 2, maxLength: 80 }
        origin: { type: string, minLength: 2, maxLength: 180 }
        destination: { type: string, minLength: 2, maxLength: 180 }
        status: { type: string, default: "Shipment created" }
        weight: { type: ["number", "null"], minimum: 0 }
        dimensions: { type: ["string", "null"], maxLength: 100 }
        packageDescription: { type: ["string", "null"], maxLength: 500 }
        estimatedDelivery: { type: ["string", "null"], format: date }
    ShipmentUpdate:
      type: object
      properties:
        customerId: { type: ["integer", "null"] }
        carrier: { type: ["string", "null"], maxLength: 100 }
        carrierTrackingNumber: { type: ["string", "null"], maxLength: 100 }
        serviceType: { type: string, minLength: 2, maxLength: 80 }
        origin: { type: string, minLength: 2, maxLength: 180 }
        destination: { type: string, minLength: 2, maxLength: 180 }
        currentLocation: { type: ["string", "null"], maxLength: 180 }
        status: { type: string, maxLength: 80 }
        weight: { type: ["number", "null"], minimum: 0 }
        dimensions: { type: ["string", "null"], maxLength: 100 }
        packageDescription: { type: ["string", "null"], maxLength: 500 }
        estimatedDelivery: { type: ["string", "null"], format: date }
        actualDelivery: { type: ["string", "null"], format: date }
    ShipmentPage:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Shipment" }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
...
    Dashboard:
      type: object
      required: [totalShipments, activeShipments, deliveredShipments, pendingQuotes, customerCount, recentEvents, statusCounts, providerConfigured]
      properties:
        totalShipments: { type: integer }
        activeShipments: { type: integer }
        deliveredShipments: { type: integer }
        pendingQuotes: { type: integer }
        customerCount: { type: integer }
        providerConfigured: { type: boolean }
        recentEvents:
          type: array
          items:
            type: object
            required: [trackingNumber, status, description, eventTimestamp, source]
            properties:
              trackingNumber: { type: string }
              status: { type: string }
              description: { type: string }
              eventTimestamp: { type: string, format: date-time }
              source: { type: string }
        statusCounts:
          type: array
          items:
            type: object
            required: [status, count]
            properties:
              status: { type: string }
              count: { type: integer }
    SyncResult:
      type: object
      required: [configured, registered, eventCount, message]
      properties:
        configured: { type: boolean }
        registered: { type: boolean }
        eventCount: { type: integer }
        message: { type: string }
...